+++ packages/plugins/type-distributor/src/ui/components/columns/s-ied/index.ts
@@ -1,0 +1,1 @@
+export { default as SIedDetails } from './s-ied-details.svelte'